+++ source/Examples/Resources/SqlAG/3-CreateAvailabilityGroupDetailed.ps1
@@ -48,6 +48,7 @@ Configuration Example
         SqlServerEndpoint 'HADREndpoint'
         {
             EndPointName         = 'HADR'
+            EndpointType         = 'DatabaseMirroring'
             Ensure               = 'Present'
             Port                 = 5022
             ServerName           = $Node.NodeName