+++ tests/TestHelpers/CommonTestHelper.psm1
@@ -252,7 +252,6 @@ function New-SQLSelfSignedCertificate
         There are build workers still on Windows Server 2012 R2 so let's
         use the alternate method of New-SelfSignedCertificate.
     #>
-    Install-Module -Name 'PSPKI' -Scope 'CurrentUser' -Force
     Import-Module -Name 'PSPKI'
 
     $newSelfSignedCertificateExParameters = @{